Fixed joysticks to use 'valuecallback' Fixed to work on MacOS 10.13 again tidyup of the Sound

+++ XCode/Project/Src/MacBridge/BeebEm-Bridging-Keyboard.cpp
@@ -655,7 +655,7 @@ extern "C" void beeb_handlejoystick(long message, long wParam, long lParam)
 			if (wParam == 3) bitpos = JOY_BUTTON3;
 			if (wParam == 4) bitpos = JOY_BUTTON4;
 
-			int oldpattern = buttons1;
+			long oldpattern = buttons1;
 
 			if (lParam==0)
 			{
